pyproject.toml
[project]
name = "agentic-ai-stock-analysis"
version = "0.1.0"
description = "agentic-ai-stock-analysis"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.5",
    "fastapi>=0.116.0",
    "uvicorn>=0.35.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "google-adk>=1.7.0",
    "gradio>=5.30.0",
    "langchain-google-genai>=2.1.5",
    "langchain-google-vertexai>=2.0.24",
    "langgraph>=0.4.5",
    "litellm",
    "serpapi==0.1.5",
    "certifi",
    "python-dotenv>=1.2.1",
]
